WIP: Move from the webpack.alias feature to a more stable vendor bundle
