--- Resources/Private/JavaScript/Host/Extensibility/API/_internal/components/index.js
@@ -1,3 +1,0 @@
-const {Components} = window['@Neos:HostPluginAPI'];
-
-module.exports = Components;
