--- Resources/Private/JavaScript/Host/Extensibility/API/_internal/immutable/index.js
@@ -1,2 +1,0 @@
-const {Immutable} = window['@Neos:HostPluginAPI'];
-module.exports = Immutable;
