--- Resources/Private/JavaScript/Host/Extensibility/API/_internal/plow-js/index.js
@@ -1,2 +1,0 @@
-const {plow} = window['@Neos:HostPluginAPI'];
-module.exports = plow;
