--- Resources/Private/JavaScript/Host/Extensibility/API/_internal/react-dom/index.js
@@ -1,5 +1,0 @@
-import API from '@host/api/index';
-
-const {_ReactDOM} = API['@Neos:RuntimeDependencies'];
-
-export default _ReactDOM;
