--- Resources/Private/JavaScript/Host/Extensibility/API/_internal/react-redux/index.js
@@ -1,2 +1,0 @@
-const {reactRedux} = window['@Neos:HostPluginAPI'];
-module.exports = reactRedux;
