--- Resources/Private/JavaScript/Host/Extensibility/API/_internal/react/index.js
@@ -1,5 +1,0 @@
-const {React} = window['@Neos:HostPluginAPI'];
-
-export default React;
-export const Component = React.Component;
-export const PropTypes = React.PropTypes;
