--- Resources/Private/JavaScript/Host/Extensibility/API/_internal/react/lib/ReactComponentWithPureRenderMixin.js
@@ -1,3 +1,0 @@
-const {React} = window['@Neos:HostPluginAPI'];
-
-export default React.ReactComponentWithPureRenderMixin;
