+++ Resources/Private/JavaScript/InspectorEditors/Image/Components/PreviewScreen/index.js
@@ -1,5 +1,5 @@
 import React, {Component, PropTypes} from 'react';
-import {Icon} from 'Components';
+import Icon from '@neos-project/react-ui-components/lib/Icon/';
 import Dropzone from 'react-dropzone';
 
 import {Thumbnail} from '../../Utils/index';
