+++ Resources/Private/JavaScript/InspectorEditors/Image/Components/Secondary/ImageCropper/AspectRatioDropDown/index.js
@@ -1,11 +1,12 @@
 import React, {Component, PropTypes} from 'react';
 import {Maybe} from 'monet';
+import IconButton from '@neos-project/react-ui-components/lib/IconButton/';
+import DropDown from '@neos-project/react-ui-components/lib/DropDown/';
 
 import {AspectRatioOption, NullAspectRatioStrategy} from '../model';
 
 import style from './style.css';
 
-import {IconButton, DropDown} from 'Components';
 
 export default class AspectRatioDropDown extends Component {
     static propTypes = {
