+++ Resources/Private/JavaScript/InspectorEditors/Image/Components/Secondary/MediaDetailsScreen/index.js
@@ -1,8 +1,9 @@
 import React, {PropTypes} from 'react';
-import SecondaryInspector from 'Host/Extensibility/API/SecondaryInspector';
 import style from './style.css';
 
 const MediaDetailsScreen = (props) => {
+    const {SecondaryInspector} = window['@Neos:HostPluginAPI'];
+
     // TODO: Media package refactoring
     window.Typo3MediaBrowserCallbacks = {
         close() {
