+++ Resources/Private/JavaScript/InspectorEditors/Image/Components/Secondary/MediaSelectionScreen/index.js
@@ -1,8 +1,9 @@
 import React, {PropTypes} from 'react';
-import SecondaryInspector from 'Host/Extensibility/API/SecondaryInspector';
 import style from './style.css';
 
 const MediaSelectionScreen = (props) => {
+    const {SecondaryInspector} = window['@Neos:HostPluginAPI'];
+
     // TODO: Media package refactoring
     window.Typo3MediaBrowserCallbacks = {
         assetChosen: assetIdentifier => {
