+++ Resources/Private/JavaScript/InspectorEditors/TextArea/index.js
@@ -1,7 +1,5 @@
 import React, {PropTypes} from '@host/react';
-import {Components} from '@host';
-
-const {TextArea} = Components;
+import TextArea from '@neos-project/react-ui-components/lib/TextArea/';
 
 const TextAreaEditor = props => {
     const {value, commit} = props;
