+++ Resources/Private/JavaScript/InspectorEditors/TextField/index.js
@@ -1,5 +1,5 @@
 import React, {PropTypes} from 'react';
-import {TextInput} from 'Components';
+import TextInput from '@neos-project/react-ui-components/lib/TextInput/';
 
 const TextField = props => {
     const {value, commit} = props;
