--- Resources/Private/JavaScript/InspectorEditors/index.js
@@ -1,1 +1,0 @@
-import './manifest';
