--- Resources/Private/JavaScript/InspectorEditors/index.spec.js
@@ -1,3 +1,0 @@
-describe('A Test', () => {
-    it('should run', () => {});
-});
