+++ Resources/Private/Templates/Backend/Guest.html
@@ -1,6 +1,8 @@
 <f:section name="guestFrameApplication">
     <script>window.neos = window.parent.neos;</script>
+    <script src="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/JavaScript/Vendor.js')}"></script>
     <script src="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/Plugins/ckeditor/ckeditor.js')}"></script>
     <script src="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/JavaScript/Guest.js')}"></script>
+    <link rel="stylesheet" href="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/Styles/Vendor.css')}">
     <link rel="stylesheet" href="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/Styles/Host.css')}">
 </f:section>
