+++ Resources/Private/Templates/Backend/Index.html
@@ -3,8 +3,10 @@
 <html>
     <head>
         <title>Neos CMS</title>
+        <script src="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/JavaScript/Vendor.js')}"></script>
         <script src="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/JavaScript/Host.js')}"></script>
         <script src="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/JavaScript/InspectorEditors.js')}"></script>
+        <link rel="stylesheet" href="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/Styles/Vendor.css')}">
         <link rel="stylesheet" href="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/Styles/Host.css')}">
         <link rel="stylesheet" href="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/Styles/HostOnlyStyles.css')}">
         <link rel="stylesheet" href="{f:uri.resource(path: 'resource://Neos.Neos.Ui/Public/Styles/InspectorEditors.css')}">
